Refactored entity names
