+++ pydash/pydash_app/dashboard/services/seeding.py
@@ -4,7 +4,7 @@ to make it easier to test code in development and staging environments.
 """
 
 
-from pydash_app.dashboard.dashboard import Dashboard
+from pydash_app.dashboard.entity import Dashboard
 import pydash_app.dashboard.repository as repository
 import pydash_app.user.repository as user_repository
