+++ pydash/pydash_web/controller/dashboards.py
@@ -6,7 +6,7 @@ Currently only returns static mock data.
 
 from flask import jsonify
 from flask_login import current_user
-from pydash_app.dashboard.dashboard import DashboardState
+from pydash_app.dashboard.entity import DashboardState
 
 import pydash_app.dashboard
 import pydash_logger
